[frontend] feat(chaining): action selection
