+++ openaev-front/src/admin/components/scenarios/scenario/logic/ScenarioLogic.tsx
@@ -8,7 +8,7 @@ import Logic from '../../../chaining/logic/Logic';
 const ScenarioLogic = () => {
   const { scenarioId } = useParams() as { scenarioId: Scenario['scenario_id'] };
   const { scenario } = useHelper((helper: ScenariosHelper) => ({ scenario: helper.getScenario(scenarioId) }));
-  return <Logic workflowId={scenario?.scenario_workflow_id} />;
+  return <Logic workflowId={scenario?.scenario_workflow_id} context="scenario" />;
 };
 
 export default ScenarioLogic;
